refator: renames page component and implents new content components
Relates# 36
