+++ src/components/page-components/ProjectDetails/CurrentProjectPage.css
@@ -1,0 +1,30 @@
+.project-details {
+  background-color: white;
+  animation: slideIn 0.5s ease-out;
+}
+
+.project-details-enter {
+  z-index: 3;
+}
+
+.project-details-enter-active {
+  z-index: 3;
+}
+
+.project-details-exit {
+  z-index: 1;
+}
+
+.project-details-exit-active {
+  z-index: 1;
+}
+
+@keyframes slideIn {
+  0% {
+    transform: translateX(200);
+  }
+
+  100% {
+    transform: translateX(0);
+  }
+}
