--- src/components/page-components/ProjectDetails/ProjectDetails.css
@@ -1,92 +1,0 @@
-.project-details {
-  background-color: white;
-  animation: slideIn 0.5s ease-out;
-}
-
-.project-details-enter {
-  z-index: 3;
-}
-
-.project-details-enter-active {
-  z-index: 3;
-}
-
-.project-details-exit {
-  z-index: 1;
-}
-
-.project-details-exit-active {
-  z-index: 1;
-}
-
-@keyframes slideIn {
-  0% {
-    transform: translateX(200);
-  }
-
-  100% {
-    transform: translateX(0);
-  }
-}
-
-.project-details img {
-  height: 14.3rem;
-  width: 100%;
-  object-fit: cover;
-}
-
-.project-details__container {
-  grid-column: 1/3;
-  margin-top: -1.25rem;
-}
-
-.project-details-description {
-  width: 340px;
-  padding: 1.25rem;
-}
-
-.project-details-description span {
-  margin-right: 0.625rem;
-}
-
-.project-details-description button {
-  margin-top: 1.25rem;
-  font-size: 0.938rem;
-}
-
-.read-more-content {
-  display: grid;
-  grid-template-rows: 0fr;
-  transition: grid-template-rows 500ms;
-  margin-top: 0.625rem;
-}
-
-.read-more-content div {
-  overflow: hidden;
-}
-
-.read-more__icon {
-  transition: transform 500ms;
-}
-
-.read-more__icon[aria-expanded="true"] {
-  transform: rotate(90deg);
-}
-
-.read-more-content[aria-hidden="false"] {
-  grid-template-rows: 1fr;
-}
-
-.project-curator-details {
-  grid-column: 1/3;
-  width: 340px;
-}
-
-.project-upcoming-events {
-  gap: 2rem;
-}
-
-.project-upcoming-events h4 {
-  font-weight: 400;
-  font-size: 0.9rem;
-}
